@@ -2,13 +2,12 @@ package main
 
 import (
 	"github.com/spf13/cobra"
-	"github.com/whitekid/cobrax"
 
 	"reader/db"
 )
 
 func init() {
-	cobrax.Add(rootCmd, &cobra.Command{
+	rootCmd.AddCommand(&cobra.Command{
 		Use:   "migrate",
 		Short: "migrate database schema",
 		RunE: func(cmd *cobra.Command, args []string) error {
@@ -18,5 +17,5 @@ func init() {
 
 			return nil
 		},
-	}, nil, nil)
+	})
 }
