@@ -9,9 +9,8 @@ import (
 	"github.com/go-testfixtures/testfixtures/v3"
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
-	"github.com/whitekid/cobrax"
 	"github.com/whitekid/goxp/log"
-	"github.com/whitekid/goxp/request"
+	"github.com/whitekid/goxp/requests"
 	"gopkg.in/yaml.v3"
 
 	"reader/db"
@@ -23,7 +22,7 @@ func init() {
 		Use: "fixtures",
 	}
 
-	cobrax.Add(cmd, &cobra.Command{
+	cmd.AddCommand(&cobra.Command{
 		Use:   "setup",
 		Short: "setup fixture database",
 		RunE: func(cmd *cobra.Command, args []string) error {
@@ -53,7 +52,7 @@ func init() {
 
 			for _, url := range fixture.Urls {
 				log.Infof("fetching %s...", url)
-				resp, err := request.Get(url).FollowRedirect(true).Do(cmd.Context())
+				resp, err := requests.Get(url).FollowRedirect(true).Do(cmd.Context())
 				if err != nil {
 					return nil
 				}
@@ -76,7 +75,7 @@ func init() {
 
 			return db.DumpFixture()
 		},
-	}, nil, nil)
+	})
 
 	cmd.AddCommand(&cobra.Command{
 		Use: "dump database",
